# src/acm-metamodel/PhysicalProcess.js
var ACMElement = require("./ACMElement.js");

class PhysicalProcess extends ACMElement {
    constructor(id, name,x,y) {
        super(id, name,x,y);
        }
    
}

module.exports = PhysicalProcess;